device-emulator/requirements.txt
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
six==1.17.0
